
\name{RefTMB}
\alias{RefTMB}

\title{
    Example Data
}
\description{ 
  The pan-cancer repertoire of reference signatures and 
  the reference TMB (Tumor mutation burden) signature profiles 
  of SBS (Single base substitutions) and
  DBS (Double base substitutions).
}
\details{
  This file consists of the list \code{RefTMB} with the following objects: 
  \itemize{
    \item{TMB_SBS_v3.2} : The version 3.2 reference SBS TMB signature profiles of size 96 by 76.
    \item{TMB_DBS_v3.2} : The version 3.2 reference DBS TMB signature profiles of size 78 by 11.
    \item{TMB_SBS_v3.4} : The version 3.4 reference SBS TMB signature profiles of size 96 by 85.
    \item{TMB_DBS_v3.4} : The version 3.4 reference DBS TMB signature profiles of size 78 by 20.
    \item{SBS_refSigs} : The list of cancer specific SBS signature names.
    \item{DBS_refSigs} : The list of cancer specific DBS signature names.
  }
 }
%\examples{}
\keyword{data}
